+++ Core/Source/Controller/MoveEdgeTool.h
@@ -28,15 +28,15 @@ namespace TrenchBroom {
         protected:
             virtual int hitType();
             virtual std::string undoName();
-            virtual Vec3f movePosition(const Model::Brush& brush, int index);
+            virtual Vec3f movePosition(const Model::Brush& brush, size_t index);
             virtual const Vec4f& handleColor();
             virtual const Vec4f& hiddenHandleColor();
             virtual const Vec4f& selectedHandleColor();
             virtual const Vec4f& hiddenSelectedHandleColor();
-            virtual Model::MoveResult performMove(Model::Brush& brush, int index, const Vec3f& delta);
+            virtual Model::MoveResult performMove(Model::Brush& brush, size_t index, const Vec3f& delta);
             
             virtual void updateHandleFigure(Renderer::HandleFigure& handleFigure);
-            virtual void updateSelectedHandleFigures(Renderer::HandleFigure& handleFigure, Renderer::PointGuideFigure& guideFigure, const Model::Brush& brush, int index);
+            virtual void updateSelectedHandleFigures(Renderer::HandleFigure& handleFigure, Renderer::PointGuideFigure& guideFigure, const Model::Brush& brush, size_t index);
         public:
             MoveEdgeTool(Controller::Editor& editor) : VertexTool(editor) {}
             virtual ~MoveEdgeTool() {}
